Refactor tests folder structure
