+++ activiti-cloud-acceptance-tests-playwright/playwright.config.ts
@@ -63,15 +63,11 @@ export default defineConfig({
       ...serial,
     },
     { name: 'smoke', grep: /@smoke/ },
-    { name: 'identity', testMatch: '**/identity-adapter.spec.ts' },
-    {
-      name: 'security',
-      testMatch: ['**/hruser-security-policies.spec.ts', '**/hradmin-security-policies.spec.ts'],
-    },
-    { name: 'process', testMatch: '**/process-instance-actions.spec.ts' },
-    { name: 'runtime', testMatch: '**/runtime/**/*.spec.ts' },
-    { name: 'runtime-process-instance', testMatch: '**/runtime/process-instance*.spec.ts' },
-    { name: 'runtime-tasks', testMatch: '**/runtime/task*.spec.ts' },
+    { name: 'identity', testDir: './tests/identity' },
+    { name: 'security', testDir: './tests/security' },
+    { name: 'runtime-process', testDir: './tests/runtime-process' },
+    { name: 'runtime-task', testDir: './tests/runtime-task' },
+    { name: 'runtime', testDir: './tests/runtime' },
   ],
   maxFailures: isCi ? 10 : undefined,
