+++ activiti-cloud-acceptance-tests-playwright/tests/runtime-process/process-variable-mapping-types.spec.ts
