+++ package.json
@@ -24,10 +24,9 @@
     "test:debug": "npm run pw -- $npm_package_config_pw_suite --debug",
     "test:identity": "npm run pw -- --project=identity",
     "test:security": "npm run pw -- --project=security",
-    "test:process": "npm run pw -- --project=process",
     "test:runtime": "npm run pw -- --project=runtime",
-    "test:runtime:process-instance": "npm run pw -- --project=runtime-process-instance",
-    "test:runtime:tasks": "npm run pw -- --project=runtime-tasks",
+    "test:runtime:process": "npm run pw -- --project=runtime-process",
+    "test:runtime:task": "npm run pw -- --project=runtime-task",
     "report": "playwright show-report",
     "install:browsers": "playwright install",
     "install:deps": "playwright install-deps",
